backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
mistralai==0.1.6
pandas==2.2.1
numpy==1.26.4
pydantic==2.6.4
cors==1.0.1
